[IMP] make remove migration folder optional
Using odoo upgrade service and skiping versions, it's useful for previous migration scripts to remain there.

Maybe this is not ok for OCA repositories but can be used on custom repositories.
