+++ odoo_module_migrate/migration_scripts/migrate_allways.py
@@ -2,8 +2,6 @@
 # @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-import os
-import subprocess
 from odoo_module_migrate.base_migration_script import BaseMigrationScript
 
 _TEXT_REPLACES = {
@@ -22,20 +20,8 @@ def set_module_installable(**kwargs):
         manifest_path, {old_term: new_term}, "Set module installable")
 
 
-def remove_migration_folder(**kwargs):
-    logger = kwargs['logger']
-    module_path = kwargs['module_path']
-    migration_path_folder = os.path.join(module_path, 'migrations')
-    if os.path.exists(migration_path_folder):
-        logger.info("Removing 'migrations' folder")
-        subprocess.check_output(
-            "rm -r %s" % migration_path_folder, shell=True
-        )
-
-
 class MigrationScript(BaseMigrationScript):
     _TEXT_REPLACES = _TEXT_REPLACES
     _GLOBAL_FUNCTIONS = [
-        remove_migration_folder,
         set_module_installable,
     ]
